+++ crates/octokit-webhooks/src/properties/reactions.rs
@@ -3,15 +3,15 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize)]
 pub struct Reactions {
   pub url: String,
-  pub total_count: u32,
+  pub total_count: u64,
   #[serde(rename = "+1")]
-  pub positive: u32,
+  pub positive: u64,
   #[serde(rename = "-1")]
-  pub negative: u32,
-  pub laugh: u32,
-  pub hooray: u32,
-  pub confused: u32,
-  pub heart: u32,
-  pub rocket: u32,
-  pub eyes: u32,
+  pub negative: u64,
+  pub laugh: u64,
+  pub hooray: u64,
+  pub confused: u64,
+  pub heart: u64,
+  pub rocket: u64,
+  pub eyes: u64,
 }
